[FEATURE] #256 : 쪽지 홈 내 ViewModel, State Home 네이밍 추가 + 쪽지 메인에서 사용하는 ViewModel 구현
