+++ feature/message/src/main/kotlin/com/bff/wespot/message/screen/MessageScreen.kt
@@ -7,16 +7,15 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.bff.wespot.designsystem.component.indicator.WSHomeTabRow
 import com.bff.wespot.message.R
 import com.bff.wespot.message.common.HOME_SCREEN_INDEX
 import com.bff.wespot.message.common.STORAGE_SCREEN_INDEX
 import com.bff.wespot.message.screen.storage.MessageStorageScreen
-import com.bff.wespot.message.state.MessageAction
 import com.bff.wespot.message.viewmodel.MessageViewModel
 import com.bff.wespot.message.viewmodel.SendViewModel
 import com.bff.wespot.model.common.RestrictionArg
@@ -24,7 +23,6 @@ import com.bff.wespot.model.notification.NotificationType
 import com.bff.wespot.ui.model.ToastState
 import com.ramcosta.composedestinations.annotation.Destination
 import kotlinx.collections.immutable.persistentListOf
-import org.orbitmvi.orbit.compose.collectAsState
 
 interface MessageNavigator {
     fun navigateUp()
@@ -46,12 +44,12 @@ internal fun MessageScreen(
     showToast: (ToastState) -> Unit,
     restricted: RestrictionArg,
 ) {
+    val selectedTabIndex = viewModel.selectedTabIndex.collectAsStateWithLifecycle().value
+
     val tabList = persistentListOf(
         stringResource(R.string.message_home_screen),
         stringResource(R.string.message_storage_screen),
     )
-    val state by viewModel.collectAsState()
-    val action = viewModel::onAction
 
     Scaffold {
         Column(
@@ -60,21 +58,20 @@ internal fun MessageScreen(
                 .padding(it),
         ) {
             WSHomeTabRow(
-                selectedTabIndex = state.selectedTabIndex,
+                selectedTabIndex = selectedTabIndex,
                 tabList = tabList,
-                onTabSelected = { action(MessageAction.OnTabSelected(it)) },
+                onTabSelected = { viewModel.updateTabIndex(it) },
             )
 
             Crossfade(
-                targetState = state.selectedTabIndex,
+                targetState = selectedTabIndex,
                 label = stringResource(R.string.message_screen_crossfade),
             ) { page ->
                 when (page) {
                     HOME_SCREEN_INDEX -> {
                         MessageHomeScreen(
-                            viewModel = viewModel,
                             navigateToMessageStorageScreen = {
-                                action(MessageAction.OnTabSelected(STORAGE_SCREEN_INDEX))
+                                viewModel.updateTabIndex(STORAGE_SCREEN_INDEX)
                             },
                             navigateToReceiverSelectionScreen = {
                                 messageNavigator.navigateReceiverSelectionScreen()
@@ -98,7 +95,7 @@ internal fun MessageScreen(
     LaunchedEffect(Unit) {
         when (navArgs.type) {
             NotificationType.MESSAGE_RECEIVED, NotificationType.MESSAGE_SENT -> {
-                action(MessageAction.OnTabSelected(STORAGE_SCREEN_INDEX))
+                viewModel.updateTabIndex(STORAGE_SCREEN_INDEX)
             }
             else -> { }
         }
