--- feature/message/src/main/kotlin/com/bff/wespot/message/state/MessageAction.kt
@@ -1,7 +1,0 @@
-package com.bff.wespot.message.state
-
-sealed class MessageAction {
-    data class OnTabSelected(val index: Int) : MessageAction()
-    data object OnLifecycleStart : MessageAction()
-    data object OnLifecycleStop : MessageAction()
-}
