--- feature/message/src/main/kotlin/com/bff/wespot/message/state/MessageSideEffect.kt
@@ -1,3 +1,0 @@
-package com.bff.wespot.message.state
-
-sealed class MessageSideEffect
