+++ feature/message/src/main/kotlin/com/bff/wespot/message/state/home/MessageHomeAction.kt
@@ -1,0 +1,7 @@
+package com.bff.wespot.message.state.home
+
+sealed class MessageHomeAction {
+    data object OnScreenEntered : MessageHomeAction()
+    data object OnLifecycleStart : MessageHomeAction()
+    data object OnLifecycleStop : MessageHomeAction()
+}
