+++ feature/message/src/main/kotlin/com/bff/wespot/message/state/home/MessageHomeSideEffect.kt
@@ -1,0 +1,3 @@
+package com.bff.wespot.message.state.home
+
+sealed class MessageHomeSideEffect
