+++ feature/message/src/main/kotlin/com/bff/wespot/message/state/home/MessageHomeUiState.kt
@@ -1,10 +1,9 @@
-package com.bff.wespot.message.state
+package com.bff.wespot.message.state.home
 
 import com.bff.wespot.model.message.response.MessageStatus
 import com.bff.wespot.model.user.response.Profile
 
-data class MessageUiState(
-    val selectedTabIndex: Int = 0,
+data class MessageHomeUiState(
     val messageStatus: MessageStatus = MessageStatus(),
     val profile: Profile = Profile(),
     val isLoading: Boolean = false,
